Svelte native references
Instead of fishing for the element that should be rendered to the PDF
with JavaScript native functions, I am now using the Svelte native way.

Is it better? I think so. Using JavaScript native functions feels more
prone to error. Also this way it is more clearly visible how the
components are connected with eachother.